Add --print-progress flag to show puller progress
Add minimal prints to indicate puller progress if --print-progress
flag is passed to fast puller.  The intention is that the bazel
container_pull workspace rules could eventually output something,
instead of bazel getting seemingly stuck for 10 minutes when
pulling an image that is many gigabytes in size.

We are intentionally not using the logging facility and the existing
--stderrthreshold argument to display these prints, because that
will produce log-formatted output that looks too detailed when inlined
with other bazel output.  The intention is to show user-friendly
progress instead:

   Downloading from gcr.io/tensorflow/tensorflow:latest (layer 1/12)
   Downloading from gcr.io/tensorflow/tensorflow:latest (layer 2/12)
   Downloading from gcr.io/tensorflow/tensorflow:latest (layer 3/12)
   Downloading from gcr.io/tensorflow/tensorflow:latest (layer 4/12)
   Downloading from gcr.io/tensorflow/tensorflow:latest (layer 5/12)
   ...